Refactor location
